+++ src/app/community/_components/PostCard.tsx
@@ -17,7 +17,7 @@ import { communityPopOverOption } from '@/libs/communityPopOverOption';
 import AuthorCard from './AuthorCard';
 import PostCardDetailModal from './PostCardDetailModal/PostCardDetailModal';
 import { PostInteractions } from './PostInteractions';
-import DetailModalSkeleton from './PostCardDetailModal/DetailModalSkeleton';
+import DetailModalSkeleton from './PostCardDetailModal/ModalSkeleton';
 import ErrorFallbackDetailModal from './PostCardDetailModal/ErrorFallbackDetailModal';
 
 import styles from './PostCard.module.scss';
@@ -33,7 +33,7 @@ export default function PostCard({ cardData, isMine }: PostCardProps) {
   const queryClient = useQueryClient();
 
   const [isPostModalOpen, setIsPostModalOpen] = useState(false);
-  const [isPopupOpen, setIsPopupOpen] = useState(false);
+  const [isPopOverOpen, setIsPopOverOpen] = useState(false);
   const [isEditModalOpen, setIsEditModalOpen] = useState(false);
 
   const { id, nickName, updateAt, title, thumbnail, likeCount, commentCount, userImage, isLiked } = cardData;
@@ -42,7 +42,7 @@ export default function PostCard({ cardData, isMine }: PostCardProps) {
   const timeToString = calculateTimeDifference(ApdatedDate);
 
   const handleClickPopOver = () => {
-    setIsPopupOpen((prevIsOpen) => !prevIsOpen);
+    setIsPopOverOpen((prevIsOpen) => !prevIsOpen);
   };
 
   const { refetch, data: detailData } = useQuery({
@@ -58,7 +58,7 @@ export default function PostCard({ cardData, isMine }: PostCardProps) {
   }, [isEditModalOpen, refetch]);
 
   const handleClosePopOver = () => {
-    setIsPopupOpen(false);
+    setIsPopOverOpen(false);
   };
 
   const handleClickPostModal = () => {
@@ -111,7 +111,7 @@ export default function PostCard({ cardData, isMine }: PostCardProps) {
           userImage={userImage}
           onClickPopOver={handleClickPopOver}
           onClosePopOver={handleClosePopOver}
-          isOpenPopOver={isPopupOpen}
+          isOpenPopOver={isPopOverOpen}
           popOverOptions={communityPopOverOption({
             isMine,
             onClickDelete: handleClickDelete,
