+++ src/app/community/_components/PostCardDetailModal/ModalSkeleton.module.scss
