+++ src/app/community/_components/PostCardDetailModal/ModalSkeleton.tsx
@@ -1,10 +1,10 @@
 import classNames from 'classnames/bind';
 import LogoLoading from '@/components/LogoLoading/LogoLoading';
-import styles from './DetailModalSkeleton.module.scss';
+import styles from './ModalSkeleton.module.scss';
 
 const cn = classNames.bind(styles);
 
-export default function DetailModalSkeleton() {
+export default function ModalSkeleton() {
   return (
     <div className={cn('container')}>
       <LogoLoading />
